requirements.txt
Django==4.0.6
django-import-export==2.8.0
djangorestframework==3.13.1
django-search-admin-autocomplete==0.2.1
django_filter==22.1
